Set up recursive bi-partition tree

recursive_biTree returns a list with the number of communities, tree path of each node, a list of node labels

recursive_biTree(
  A,
  cluster.labels.list,
  ncluster,
  cluster.labels,
  n.min,
  method
)

Arguments

A
It is the consensus graph (the long format of consensus matrix from consensus_build (node to node connection weight))

cluster.labels.list
To store a list of labels for each cluster during the process

ncluster
To store the number of clusters during the process

cluster.labels
A vector of labels for all the input nodes, the order should be consistent with the order of A

n.min
The minimum number of nodes for further partition

method
The partition method in spectral clustering

Value

a list with the number of communities, tree path of each node,a list of community labels for each community

Examples

library(dplyr,quietly=TRUE,warn.conflicts=FALSE)
library(igraph,quietly=TRUE,warn.conflicts=FALSE)

g <- graph_from_data_frame(consensus_dat,directed = F,vertices = NULL)
g <- simplify(g)
adj <- get.adjacency(g,type = "both",attr = "weight")

id_label <- bisbmresults[[1]] %>%
         filter(type=="id") %>%
         filter(level==0) %>%
         dplyr::select(node) %>%
         arrange(node)
id_label <- id_label$node

cluster.result <- recursive_biTree(adj,cluster.labels.list=list(),
                                   ncluster=0, cluster.labels=id_label,
                                   n.min=10,method="kmeans")

head(cluster.result[[1]])

#> [[1]]
#>   [1] "R100" "R101" "R122" "R124" "R125" "R146" "R149" "R151" "R152" "R155"
#>  [11] "R156" "R157" "R158" "R166" "R167" "R172" "R177" "R179" "R180" "R182"
#>  [21] "R183" "R186" "R188" "R190" "R193" "R195" "R196" "R199" "R200" "R201"
#>  [31] "R202" "R203" "R204" "R205" "R206" "R210" "R211" "R212" "R214" "R215"
#>  [41] "R218" "R219" "R220" "R222" "R223" "R224" "R225" "R227" "R228" "R230"
#>  [51] "R231" "R235" "R237" "R238" "R239" "R240" "R242" "R243" "R244" "R245"
#>  [61] "R246" "R248" "R249" "R250" "R251" "R253" "R254" "R255" "R256" "R259"
#>  [71] "R260" "R262" "R263" "R264" "R266" "R267" "R268" "R269" "R273" "R274"
#>  [81] "R278" "R280" "R282" "R284" "R286" "R287" "R288" "R289" "R290" "R291"
#>  [91] "R292" "R346" "R347" "R348" "R349" "R368" "R378" "R45"  "R58"  "R89" 
#> [101] "R90" 
#> 
#> [[2]]
#>  [1] "R118" "R12"  "R129" "R135" "R14"  "R15"  "R153" "R16"  "R18"  "R22" 
#> [11] "R226" "R23"  "R24"  "R257" "R27"  "R277" "R28"  "R281" "R29"  "R293"
#> [21] "R298" "R3"   "R30"  "R306" "R309" "R312" "R314" "R315" "R317" "R321"
#> [31] "R324" "R326" "R327" "R33"  "R332" "R335" "R35"  "R364" "R37"  "R38" 
#> [41] "R51"  "R54"  "R6"   "R62"  "R71"  "R79"  "R8"   "R97" 
#> 
#> [[3]]
#>  [1] "R10"  "R11"  "R123" "R133" "R148" "R150" "R161" "R164" "R174" "R184"
#> [11] "R19"  "R191" "R192" "R197" "R198" "R20"  "R207" "R208" "R209" "R21" 
#> [21] "R217" "R229" "R252" "R258" "R265" "R270" "R275" "R276" "R279" "R283"
#> [31] "R301" "R302" "R303" "R308" "R311" "R328" "R329" "R338" "R342" "R362"
#> [41] "R365" "R367" "R39"  "R4"   "R40"  "R42"  "R48"  "R66"  "R68"  "R9"  
#> 
#> [[4]]
#>  [1] "R17"  "R181" "R187" "R189" "R2"   "R213" "R25"  "R26"  "R32"  "R34" 
#> [11] "R345" "R369" "R44"  "R47"  "R50"  "R7"  
#> 
#> [[5]]
#>  [1] "R13"  "R131" "R132" "R136" "R142" "R144" "R46"  "R77"  "R78"  "R81" 
#> [11] "R82"  "R86"  "R91"  "R92"  "R93"  "R94"  "R95"  "R99" 
#> 
#> [[6]]
#>  [1] "R130" "R138" "R139" "R140" "R221" "R241" "R271" "R272" "R299" "R31" 
#> [11] "R319" "R322" "R330" "R350" "R41"  "R5"   "R75"  "R80" 
#> 

head(cluster.result[[2]])

#> [1] 8

head(cluster.result[[3]])

#> [1] ""       "l/"     "l/R100" "l/R101" "l/R122" "l/R124"